+++ src/data/view/unit_port.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 #[cfg(feature = "swagger")]
 use utoipa::ToSchema;
 
-/// Unit object exchanged by the PopRaKo translation port.
+/// Unit object exchanged by the `PopRaKo` translation port.
 #[derive(Debug, Clone, Deserialize, Serialize)]
 #[cfg_attr(feature = "swagger", derive(ToSchema))]
 pub struct UnitTranslationPortView {
@@ -13,12 +13,12 @@ pub struct UnitTranslationPortView {
     /// Unit identifier from the exporting chapter.
     pub unit_id: String,
     /// Ordinal index of the unit within its page.
-    pub unit_index: i32,
+    pub unit_index: usize,
 
     /// Parent page identifier from the exporting chapter.
     pub page_id: String,
     /// Ordinal index of the page within its chapter.
-    pub page_index: i32,
+    pub page_index: usize,
 
     /// Horizontal coordinate of the unit bounding box on the page.
     pub x_coord: f64,
